split JSON(De)serializer into multiple files
