+++ src/nodesystem/nodemodel.cpp
@@ -9,8 +9,9 @@
 #include "nodesystem/port.h"
 #include "scene/mailbox.h"
 #include "scene/scene.h"
-#include "serializers/jsonserializer.h"
-#include "serializers/abstractdeserializer.h"
+#include "serializers/json/jsonserializer.h"
+#include "serializers/json/jsondeserializer.h"
+
 
 namespace
 {
