+++ src/objects/object.cpp
@@ -3,29 +3,30 @@
 #include "common.h"
 #include "logging.h"
 #include "objects/pathobject.h"
-#include "path/path.h"
 #include "path/lib2geomadapter.h"
+#include "path/path.h"
 #include "path/pathvector.h"
 #include "properties/boolproperty.h"
 #include "properties/floatproperty.h"
 #include "properties/floatvectorproperty.h"
 #include "properties/integerproperty.h"
 #include "properties/integervectorproperty.h"
-#include "properties/propertygroups/markerproperties.h"
 #include "properties/optionproperty.h"
+#include "properties/propertygroups/markerproperties.h"
 #include "properties/referenceproperty.h"
 #include "properties/stringproperty.h"
-#include "renderers/style.h"
+#include "removeif.h"
 #include "renderers/painter.h"
 #include "renderers/painteroptions.h"
+#include "renderers/style.h"
 #include "scene/contextes.h"
 #include "scene/mailbox.h"
 #include "scene/objecttree.h"
 #include "scene/scene.h"
-#include "serializers/jsonserializer.h"
+#include "serializers/json/jsonserializer.h"
+#include "serializers/abstractdeserializer.h"
 #include "tags/styletag.h"
 #include "tags/tag.h"
-#include "removeif.h"
 
 #include <QObject>
 #include <QPen>
