+++ src/scene/scene.cpp
@@ -21,7 +21,8 @@
 #include "properties/referenceproperty.h"
 #include "properties/stringproperty.h"
 #include "renderers/style.h"
-#include "serializers/jsonserializer.h"
+#include "serializers/json/jsonserializer.h"
+#include "serializers/json/jsondeserializer.h"
 #include "tags/nodestag.h"
 #include "tags/tag.h"
 #include "tools/selectobjectstool.h"
