+++ src/serializers/json/CMakeLists.txt
@@ -1,0 +1,11 @@
+target_sources(libommpfritt PRIVATE
+  common.h
+  jsondeserializer.cpp
+  jsondeserializer.h
+  jsondeserializerworker.cpp
+  jsondeserializerworker.h
+  jsonserializer.cpp
+  jsonserializer.h
+  jsonserializerworker.cpp
+  jsonserializerworker.h
+)
