+++ src/serializers/json/common.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace omm::serialization
+{
+
+constexpr auto inf_value = "inf";
+constexpr auto neg_inf_value = "-inf";
+
+}  // namespace omm::serialization
