+++ src/serializers/json/jsonserializer.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "external/json_fwd.hpp"
+#include "serializers/abstractserializer.h"
+
+namespace omm::serialization
+{
+
+class JSONSerializer : public AbstractSerializer
+{
+public:
+  explicit JSONSerializer(nlohmann::json& json);
+  std::unique_ptr<SerializerWorker> sub(const std::string& key) override;
+private:
+  nlohmann::json& m_json;
+};
+
+}  // namespace omm
